Further refactored location of file managers
