+++ EonZeNx.ApexTools.Core/EonZeNx.ApexTools.Core.csproj
@@ -4,4 +4,8 @@
         <TargetFramework>net5.0</TargetFramework>
     </PropertyGroup>
 
+    <ItemGroup>
+      <Folder Include="Models\Managers" />
+    </ItemGroup>
+
 </Project>
